perf(test): move the timezone suite out of the parallel unit runner
Three calendar timezone tests were timing out at exactly the 20s budget,
and not only in CI -- a full local `bun run test` reproduced it. They are
not slow tests: run alone they finish in 280ms.

`TZ` is read once per process, so proving the week-key walk under four
zones genuinely needs four subprocesses. `test:unit` runs `--parallel`,
where those blocking `spawnSync` calls compete with the rest of the suite
for the same cores, and Bun startup -- not the assertion -- is the entire
cost. Batching the spawns first (12 startups down to 4, one per zone
rather than one per assertion) helped locally but still crossed the
budget under load, which is what identified the runner rather than the
count as the problem.

A test that spawns four Bun processes is an integration test.
`test:integration` runs serially, so the same work finishes in 98ms with
no budget pressure and no assertion weakened. The spawn batching is kept:
it is a real reduction either way.

Full suite: 5893 pass, 0 fail, where the same run previously failed two.
